+++ core/src/model/Empty.java
@@ -1,7 +1,7 @@
 package model;
 
 public class Empty extends StaticGameElement {
-    public Empty(int x, int y) {
+    public Empty(float x, float y) {
         super(x, y);
     }
 }
