+++ core/src/model/PlayerTank.java
@@ -1,7 +1,7 @@
 package model;
 
 public class PlayerTank extends GameElement {
-    public PlayerTank(int x, int y) {
+    public PlayerTank(float x, float y) {
         super(x, y, 2);
     }
 }
